refactor: relocate app env to config from utils
