+++ apps/web/src/features/auth/ui/kakao-login-button/KakaoLoginButton.tsx
@@ -8,11 +8,10 @@ import { Flex, KakaoTalkBlackLogo, mergeStyles } from '@causw/cds';
 
 import { kakaoNativeLogin } from '@/features/auth/api';
 
-import { BASE_URL, isLocal } from '@/shared/config';
+import { BASE_URL, isLocal, isMobile } from '@/shared/config';
 import { requestNativeSocialLogin } from '@/shared/lib/capacitor';
 import { toast } from '@/shared/model';
-import { extractErrorMessage, isMobile } from '@/shared/utils';
-
+import { extractErrorMessage } from '@/shared/utils';
 type KakaoLoginButtonProps = ComponentProps<'button'>;
 
 export const KakaoLoginButton = ({
