+++ apps/web/src/shared/utils/index.ts
@@ -1,3 +1,2 @@
 ﻿export * from './storage';
-export * from './env';
 export * from './api';
